Updates for doctext for enum and single-arg formatting
git-svn-id: https://svn.mcs.anl.gov/repos/sowing/trunk@289 43f3db66-2a03-4d6b-8051-32b12e86fc3a

+++ src/doctext/incfiles.cc
@@ -70,7 +70,7 @@ int SaveIncludeFile( InStream *ins, char *matchstring )
     char ch;
     int  rc = 0, state, base_state, i;
 
-    state = strlen( matchstring ) - 1;
+    state = (int)strlen( matchstring ) - 1;
     base_state = state;
     while (state >= 0 && !ins->GetChar( &ch )) {
     	// Check for end of definition
